splitting solver.h into utils and actual solver
